# Conflicts:
#	Assets/Fonts/RIDIBatang SDF.asset
#	Assets/Scenes/Test/AS/AS_RandomMap.unity

+++ Assets/Test/2ENO/ConsumeManager/SaveData/ConsumableSaveData_0.cs
@@ -12,6 +12,6 @@ public class ConsumableSaveData_0 : SaveDataBase
 
     public int date;
     public float tiredness;
-    public int hunger;
+    public float hunger;
     public float bonfireTime;
 }
